--- a/pkg/server/grpc_server.go
+++ b/pkg/server/grpc_server.go
@@ -26,15 +26,19 @@ import (
 // RPCs.
 type grpcServer struct {
 	*grpc.Server
-	mode serveMode
+	serverInterceptorsInfo rpc.ServerInterceptorInfo
+	mode                   serveMode
 }
 
 func newGRPCServer(rpcCtx *rpc.Context) *grpcServer {
 	s := &grpcServer{}
 	s.mode.set(modeInitializing)
-	s.Server = rpc.NewServer(rpcCtx, rpc.WithInterceptor(func(path string) error {
-		return s.intercept(path)
-	}))
+	srv, interceptorInfo := rpc.NewServerEx(
+		rpcCtx, rpc.WithInterceptor(func(path string) error {
+			return s.intercept(path)
+		}))
+	s.Server = srv
+	s.serverInterceptorsInfo = interceptorInfo
 	return s
 }