--- a/pkg/server/server.go
+++ b/pkg/server/server.go
@@ -887,7 +887,7 @@ func (s *Server) PreStart(ctx context.Context) error {
 
 	// Connect the node as loopback handler for RPC requests to the
 	// local node.
-	s.rpcContext.SetLocalInternalServer(s.node)
+	s.rpcContext.SetLocalInternalServer(s.node, s.grpc.serverInterceptorsInfo)
 
 	// Load the TLS configuration for the HTTP server.
 	uiTLSConfig, err := s.rpcContext.GetUIServerTLSConfig()