Before this patch, the gRPC interceptors that we normally run on the
client and the server were not being run for "RPCs" to the "Internal"
sevice going to the local node. Such RPCs are done through the
localClientAdapter, which bypasses gRPC for performance reasons. Not
running these interceptors is bad - they do useful things around
creating tasks, tracing, etc. Also authentication/authorization,
although that interceptor in particular is not subject to this change.
Some of the things done by the interceptors was done separately and
redundantly in random places to account for the RPC local case.

This patch unifies the paths of local and remote RPCs by having the
localClientAdapter run server interceptors. The client interceptors
might come in a different patch. Care was taken to not introduce a
performance regression on the local path; in particular, there are no
new allocations.

As an exception, the Auth server interceptor is skipped on the local
path. As it stands, that interceptor refuses local RPCs in secure
clusters because it can't find TLS information in the ctx. I don't know
whether we should try to make this interceptor work or not.

Node.batchInternal was simplified slightly to not create a task, since
the interceptor now does it in all cases. Before, creating the task was
redundant in the non-local RPC case.

Touches #74732. That issue calls for both client and server interceptors
to be run; this patch only deals with server interceptors.

name                        old time/op    new time/op    delta
KV/Insert/Native/rows=1-32    88.0µs ± 4%    88.3µs ± 3%    ~     (p=1.000 n=5+5)
KV/Insert/SQL/rows=1-32        333µs ± 3%     322µs ± 3%    ~     (p=0.056 n=5+5)
KV/Update/Native/rows=1-32     111µs ± 3%     113µs ± 2%    ~     (p=0.095 n=5+5)
KV/Update/SQL/rows=1-32        437µs ± 4%     435µs ± 3%    ~     (p=0.841 n=5+5)
KV/Delete/Native/rows=1-32    92.2µs ±12%    88.5µs ± 5%    ~     (p=0.421 n=5+5)
KV/Delete/SQL/rows=1-32        260µs ±13%     285µs ±17%    ~     (p=0.151 n=5+5)
KV/Scan/Native/rows=1-32      18.9µs ± 2%    19.2µs ± 2%    ~     (p=0.151 n=5+5)
KV/Scan/SQL/rows=1-32          203µs ±10%     206µs ± 6%    ~     (p=0.841 n=5+5)

name                        old alloc/op   new alloc/op   delta
KV/Insert/Native/rows=1-32    15.2kB ± 0%    15.3kB ± 2%    ~     (p=1.000 n=5+5)
KV/Insert/SQL/rows=1-32       40.2kB ± 0%    40.0kB ± 1%    ~     (p=0.151 n=5+5)
KV/Update/Native/rows=1-32    20.7kB ± 0%    20.8kB ± 1%  +0.70%  (p=0.008 n=5+5)
KV/Update/SQL/rows=1-32       46.3kB ± 0%    46.6kB ± 0%  +0.56%  (p=0.032 n=5+5)
KV/Delete/Native/rows=1-32    14.8kB ± 1%    14.8kB ± 1%    ~     (p=0.841 n=5+5)
KV/Delete/SQL/rows=1-32       41.5kB ± 1%    41.9kB ± 2%    ~     (p=0.310 n=5+5)
KV/Scan/Native/rows=1-32      6.96kB ± 0%    6.95kB ± 1%    ~     (p=0.905 n=5+5)
KV/Scan/SQL/rows=1-32         18.6kB ± 0%    18.6kB ± 0%    ~     (p=0.841 n=5+5)

name                        old allocs/op  new allocs/op  delta
KV/Insert/Native/rows=1-32       118 ± 0%       119 ± 4%    ~     (p=0.556 n=4+5)
KV/Insert/SQL/rows=1-32          331 ± 0%       330 ± 0%    ~     (p=0.286 n=5+5)
KV/Update/Native/rows=1-32       164 ± 0%       164 ± 0%    ~     (all equal)
KV/Update/SQL/rows=1-32          481 ± 0%       482 ± 0%    ~     (p=0.238 n=5+5)
KV/Delete/Native/rows=1-32       118 ± 1%       117 ± 1%    ~     (p=1.000 n=5+5)
KV/Delete/SQL/rows=1-32          307 ± 1%       308 ± 0%    ~     (p=0.238 n=5+4)
KV/Scan/Native/rows=1-32        49.0 ± 0%      49.0 ± 0%    ~     (all equal)
KV/Scan/SQL/rows=1-32            229 ± 0%       229 ± 0%    ~     (p=0.556 n=4+5)

Release note: None